+++ docs/content/components/layout/tiles/tiles-complex.demo.tsx
@@ -12,7 +12,7 @@ export default () => (
     <Card p={4}>
       <Stack space={6} alignX="center">
         <Image
-          src="https://www.pokewiki.de/images/6/63/Sugimori_004.png"
+          src="https://www.pokewiki.de/images/0/09/Hauptartwork_004.png"
           alt="glumanda"
           width={200}
           height={200}
@@ -28,7 +28,7 @@ export default () => (
     <Card p={4}>
       <Stack space={6} alignX="center">
         <Image
-          src="https://www.pokewiki.de/images/7/7a/Sugimori_005.png"
+          src="https://www.pokewiki.de/images/0/08/Hauptartwork_005.png"
           alt="glutexo"
           width={200}
           height={200}
@@ -45,7 +45,7 @@ export default () => (
     <Card p={4}>
       <Stack space={6} alignX="center">
         <Image
-          src="https://www.pokewiki.de/images/9/96/Sugimori_006.png"
+          src="https://www.pokewiki.de/images/d/de/Hauptartwork_006.png"
           alt="glurak"
           width={250}
           height={200}