+++ packages/components/src/Headline/Headline.tsx
@@ -15,9 +15,15 @@ type RemovedProps = 'className' | 'level';
 export interface HeadlineProps
   extends Omit<RAC.HeadingProps, RemovedProps>,
     TextAlignProp {
+  /**
+   * Set the color of the headline.
+   */
   color?: string;
   variant?: string;
   size?: string;
+  /**
+   * Set a different level from theme, values are from 1 - 6.
+   */
   level?: '1' | '2' | '3' | '4' | '5' | '6' | RAC.HeadingProps['level'];
 }